+++ components/collector/tests/source_collectors/gitlab/test_failed_jobs.py
@@ -1,9 +1,9 @@
 """Unit tests for the GitLab jobs collectors."""
 
-from .base import CommonGitLabJobsTestsMixin, GitLabTestCase
+from .base import CommonGitLabJobsTestsMixin, GitLabJobsTestCase
 
 
-class GitLabFailedJobsTest(CommonGitLabJobsTestsMixin, GitLabTestCase):
+class GitLabFailedJobsTest(CommonGitLabJobsTestsMixin, GitLabJobsTestCase):
     """Unit tests for the GitLab failed jobs metric."""
 
     METRIC_TYPE = "failed_jobs"
@@ -11,20 +11,20 @@ class GitLabFailedJobsTest(CommonGitLabJobsTestsMixin, GitLabTestCase):
     async def test_nr_of_failed_jobs(self):
         """Test that the number of failed jobs is returned."""
         response = await self.collect(get_request_json_return_value=self.gitlab_jobs_json)
-        self.assert_measurement(response, value="2", entities=self.expected_entities)
+        self.assert_measurement(response, value="2", entities=self.expected_entities, landing_url=self.LANDING_URL)
 
     async def test_nr_of_failed_jobs_without_failed_jobs(self):
         """Test that the number of failed jobs is returned."""
         for job in self.gitlab_jobs_json:
             job["status"] = "success"
         response = await self.collect(get_request_json_return_value=self.gitlab_jobs_json)
-        self.assert_measurement(response, value="0", entities=[])
+        self.assert_measurement(response, value="0", entities=[], landing_url=self.LANDING_URL)
 
     async def test_no_jobs_in_lookback_period(self):
         """Test that the number of failed jobs is returned."""
         self.set_source_parameter("lookback_days", "3")
         response = await self.collect(get_request_json_return_value=self.gitlab_jobs_json)
-        self.assert_measurement(response, value="0", entities=[])
+        self.assert_measurement(response, value="0", entities=[], landing_url=self.LANDING_URL)
 
     async def test_ignore_previous_runs_of_jobs(self):
         """Test that previous runs of the same job are ignored."""
@@ -51,7 +51,7 @@ class GitLabFailedJobsTest(CommonGitLabJobsTestsMixin, GitLabTestCase):
             ],
         )
         response = await self.collect(get_request_json_return_value=self.gitlab_jobs_json)
-        self.assert_measurement(response, value="1", entities=self.expected_entities[-1:])
+        self.assert_measurement(response, value="1", entities=self.expected_entities[-1:], landing_url=self.LANDING_URL)
 
     async def test_private_token(self):
         """Test that the private token is used."""
@@ -64,4 +64,5 @@ class GitLabFailedJobsTest(CommonGitLabJobsTestsMixin, GitLabTestCase):
                 "https://gitlab/api/v4/projects/namespace%2Fproject/jobs?per_page=100&"
                 "scope[]=canceled&scope[]=failed&scope[]=skipped"
             ),
+            landing_url=self.LANDING_URL,
         )